Update vorto to nightly version tag
Since the vorto docker hub repo does not feature a working 0.11.x image
we'll go with the nightly image and hope nothing breaks.
We'll also change the dependency direction, since in vorto 0.11.x and
API v2 the generators won't register themselfes anymore but are
configured in the repository application. This is now also a
soft-dependency.