Migrate to vorto 0.11.3
This migration is based on the generator-runner from the vorto
and vorto-examples repos. The runner is basically an API v1 runner
with an additional API v1 adapter that maps the actual plugin to
API v2.
The actual skeleton generator plugin is also migrated to API v2, using
jsonschema and eclipseditto API v2 plugins as source.
The plugin registration is now done in the repository application
and is not part of this generator project anymore. A notice about
this is added to the readme file.